Move search box component to generics
