+++ src/components/library/work/List.jsx
@@ -5,7 +5,7 @@ import { useOutletContext, useParams, useSearchParams } from 'react-router'
 import { loadLibraryEntries } from 'actions/library'
 import ListingList from 'components/generics/listing/List'
 import Navigator from 'components/generics/Navigator'
-import SearchBox from 'components/library/SearchBox'
+import SearchBox from 'components/generics/SearchBox'
 import WorkEntry from 'components/library/work/Entry'
 import NotFound from 'components/navigation/NotFound'
 import { Status } from 'reducers/alterationsResponse'
