+++ src/style/components/generics/_index.scss
@@ -11,6 +11,7 @@
 @use 'navigator';
 @use 'notification';
 @use 'permission_text';
+@use 'searchbox';
 @use 'shapes';
 @use 'tab_bar';
 @use 'token_widget';
