+++ src/style/components/generics/_searchbox.scss
