+++ src/style/components/library/_index.scss
@@ -5,7 +5,6 @@
 //
 
 @use 'artist';
-@use 'searchbox';
 @use 'song';
 @use 'status';
 @use 'widgets';
